refactor(back): move lib/friends.ts to lib/server/friends.ts
