+++ src/lib/server/friends.ts
